move protocols/message_counter to protocols/secure_channel
