+++ src/BUILD.gn
@@ -53,7 +53,7 @@ if (chip_build_tests) {
       "${chip_root}/src/lib/support/tests",
       "${chip_root}/src/messaging/tests",
       "${chip_root}/src/protocols/bdx/tests",
-      "${chip_root}/src/protocols/message_counter/tests",
+      "${chip_root}/src/protocols/secure_channel/tests",
       "${chip_root}/src/protocols/secure_channel/tests",
       "${chip_root}/src/system/tests",
       "${chip_root}/src/transport/retransmit/tests",
