+++ src/messaging/tests/echo/common.cpp
@@ -27,12 +27,12 @@
 #include "common.h"
 #include <core/CHIPCore.h>
 #include <platform/CHIPDeviceLayer.h>
-#include <protocols/message_counter/MessageCounterManager.h>
+#include <protocols/secure_channel/MessageCounterManager.h>
 #include <support/ErrorStr.h>
 
 // The ExchangeManager global object.
 chip::Messaging::ExchangeManager gExchangeManager;
-chip::message_counter::MessageCounterManager gMessageCounterManager;
+chip::secure_channel::MessageCounterManager gMessageCounterManager;
 
 void InitializeChip(void)
 {
