--- src/protocols/message_counter/tests/BUILD.gn
@@ -1,41 +1,0 @@
-# Copyright (c) 2021 Project CHIP Authors
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-# http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-import("//build_overrides/build.gni")
-import("//build_overrides/chip.gni")
-import("//build_overrides/nlio.gni")
-import("//build_overrides/nlunit_test.gni")
-
-import("${chip_root}/build/chip/chip_test_suite.gni")
-
-chip_test_suite("tests") {
-  output_name = "libMessageCounterManagerTests"
-
-  test_sources = [ "TestMessageCounterManager.cpp" ]
-
-  cflags = [ "-Wconversion" ]
-
-  public_deps = [
-    "${chip_root}/src/inet/tests:helpers",
-    "${chip_root}/src/lib/core",
-    "${chip_root}/src/lib/support",
-    "${chip_root}/src/messaging",
-    "${chip_root}/src/messaging/tests:helpers",
-    "${chip_root}/src/protocols",
-    "${chip_root}/src/transport",
-    "${chip_root}/src/transport/raw/tests:helpers",
-    "${nlio_root}:nlio",
-    "${nlunit_test_root}:nlunit-test",
-  ]
-}
