+++ src/protocols/secure_channel/tests/BUILD.gn
@@ -10,6 +10,7 @@ chip_test_suite("tests") {
 
   test_sources = [
     "TestCASESession.cpp",
+    "TestMessageCounterManager.cpp",
     "TestPASESession.cpp",
     "TestStatusReport.cpp",
   ]
@@ -19,6 +20,7 @@ chip_test_suite("tests") {
     "${chip_root}/src/lib/core",
     "${chip_root}/src/lib/support",
     "${chip_root}/src/messaging/tests:helpers",
+    "${chip_root}/src/protocols",
     "${chip_root}/src/protocols/secure_channel",
     "${nlio_root}:nlio",
     "${nlunit_test_root}:nlunit-test",
